requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.3
asgiref==3.8.1
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
datasets==3.4.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
Django==4.2.14
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.13.3
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.43.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.3.1
future @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gast==0.5.4
gdown==5.1.0
google-pasta==0.2.0
grpcio==1.62.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
inquirerpy==0.3.4
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.9.0
joblib==1.3.2
jsonify==0.5
jupyter_client==8.4.0
jupyter_core==5.4.0
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.42.0
macholib @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.5.8
networkx==3.2.1
numba==0.59.1
numpy==1.26.1
openai==1.70.0
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
pfzy==0.3.4
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.22.0
plotly-express==0.4.1
pooch==1.8.1
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.16.1
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyJWT==2.8.0
pyparsing==3.1.1
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==25.1.1
qrcode==7.4.2
redis==5.0.4
requests==2.32.3
rich==13.7.1
scikit-learn==1.3.1
scikit-surprise==1.1.3
scipy==1.11.3
seaborn==0.13.0
selenium==4.31.0
six @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sklearn==0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spotipy==2.23.0
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.2.0
torch==2.2.2
torchaudio==2.2.2
torchsummary==1.5.1
torchvision==0.17.2
tornado==6.3.3
tqdm==4.67.1
traitlets==5.11.2
trio==0.29.0
trio-websocket==0.12.2
twilio==9.0.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2023.3
urllib3==2.2.1
wcwidth==0.2.8
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.17.0
